refactor(logging): split index.ts into internal modules
